test_web_service/test_web_service/requirements.txt
Django==3.0.5
django-nested-admin == 3.3.2
Pillow==7.0.0
djangorestframework==3.11.1


